+++ src/main/java/com/postech/restaurantes/adapter/presenter/view/RoleView.java
@@ -1,0 +1,6 @@
+package com.postech.restaurantes.adapter.presenter.view;
+
+import java.util.UUID;
+
+public record RoleView(UUID id, String name) {
+}